worker/requirements.txt
celery==5.0.5
celery-redbeat==2.0.0
redis==3.5.3
opencv-python==4.4.0.46
requests==2.24.0
Pillow~=8.0.1
PyMySQL==1.0.2
SQLAlchemy==1.3.22
